renamed add securities button component
